requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.0
Brotli==1.0.9
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cleo==1.0.0a5
crashtest==0.3.1
distlib==0.3.6
dj-database-url==1.0.0
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
dulwich==0.20.46
filelock==3.8.0
gunicorn==20.1.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
jaraco.classes==3.2.2
jsonschema==4.16.0
keyring==23.9.1
lockfile==0.12.2
more-itertools==8.14.0
msgpack==1.0.4
packaging==21.3
pexpect==4.8.0
Pillow==9.1.1
pkginfo==1.8.3
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
poetry==1.2.0
poetry-core==1.1.0
poetry-plugin-export==1.0.6
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
pylev==1.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytz==2022.1
requests==2.28.1
requests-toolbelt==0.9.1
shellingham==1.5.0
six==1.16.0
sqlparse==0.4.2
tomlkit==0.11.4
urllib3==1.26.12
virtualenv==20.16.5
webencodings==0.5.1
whitenoise==6.2.0
xattr==0.9.9
your-app==0.1.0
zipp==3.8.1
